Refactoring, NodeJS-like child process API, inter-script message passing via named channels
The inter-script message passing allows for separation of concerns by delegating operations to other scripts. When a message is passed to another script, that script is executed in its own context as usual, and may be written in a completely different language from the sender.

An example use case could look like this:

- database.js \
  Holds connections to the database. Accepts queries over the `database_query` channel.
- todos_manager.py \
  Manages the todos database. Accepts new todos over the `todos_add` channel and adds them to the database by sending an SQL query to the `database_query` channel.
- todos_web.js \
  Provides the todos REST API. Submits new todos to the `todos_add` channel.